Create and use TicketingTile
