+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Assistant/TicketAssistantTile.tsx
@@ -1,19 +1,15 @@
 import {StyleSheet, useTheme} from '@atb/theme';
-import {
-  getStaticColor,
-  getTransportationColor,
-  StaticColor,
-} from '@atb/theme/colors';
-import {TouchableOpacity, View} from 'react-native';
-import {screenReaderPause, ThemeText} from '@atb/components/text';
+
+import {View} from 'react-native';
+import {screenReaderPause} from '@atb/components/text';
 import React from 'react';
 import {TicketingTexts, useTranslation} from '@atb/translations';
 import {TicketMultiple} from '@atb/assets/svg/mono-icons/ticketing';
 import {FareProductTypeConfig} from '@atb-as/config-specs';
 import {useFirestoreConfiguration} from '@atb/configuration';
 import {isProductSellableInApp} from '@atb/reference-data/utils';
-import {BetaTag} from '@atb/components/beta-tag';
 import {useTicketingState} from '@atb/ticketing';
+import {TicketingTile} from '../TicketingTile';
 
 type TicketAssistantProps = {
   accented?: boolean;
@@ -26,21 +22,14 @@ export const TicketAssistantTile: React.FC<TicketAssistantProps> = ({
   testID,
 }) => {
   const styles = useStyles();
-  const {theme, themeName} = useTheme();
-  const color: StaticColor = accented ? 'background_accent_3' : 'background_0';
-  const themeColor = getStaticColor(themeName, color);
   const {t} = useTranslation();
+  const {theme} = useTheme();
 
-  const transportThemePrimaryColor = getTransportationColor(
-    themeName,
-    'transport_other',
-    'primary',
-  );
+  const title = t(TicketingTexts.ticketAssistantTile.title);
+  const description = t(TicketingTexts.ticketAssistantTile.description);
 
-  const transportThemeSecondaryColor = getTransportationColor(
-    themeName,
-    'transport_other',
-    'secondary',
+  const accessibilityLabel = [title, 'Beta', description].join(
+    screenReaderPause,
   );
 
   const {fareProductTypeConfigs, preassignedFareProducts} =
@@ -59,103 +48,36 @@ export const TicketAssistantTile: React.FC<TicketAssistantProps> = ({
     (config) => config.configuration.requiresLogin,
   );
 
-  const title = t(TicketingTexts.ticketAssistantTile.title);
-  const description = t(TicketingTexts.ticketAssistantTile.description);
-  const accessibilityLabel = [title, 'Beta', description].join(
-    screenReaderPause,
-  );
-
-  return (
-    <View
-      style={[
-        styles.tipsAndInformation,
-        {
-          backgroundColor: themeColor.background,
-          borderBottomColor: transportThemePrimaryColor.background,
-        },
-      ]}
-      testID={testID}
-    >
-      {requiresLoginConfig && (
-        <TouchableOpacity
+  if (!requiresLoginConfig) {
+    return <View style={{height: theme.spacings.medium}} />;
+  } else {
+    return (
+      <View style={styles.tipsAndInformation}>
+        <TicketingTile
+          accented={accented}
           onPress={() => onPress(requiresLoginConfig)}
-          accessible={true}
+          testID={testID}
+          transportColor="transport_other"
+          title={title}
+          description={description}
           accessibilityLabel={accessibilityLabel}
-          accessibilityHint={t(TicketingTexts.ticketAssistantTile.a11yHint)}
-          style={styles.spreadContent}
+          showBetaTag={true}
         >
-          <View style={styles.contentContainer}>
-            <View style={styles.titleContainer}>
-              <BetaTag style={styles.betaTag} />
-            </View>
-            <View style={styles.titleContainer}>
-              <ThemeText
-                type="body__secondary--bold"
-                accessibilityLabel={t(TicketingTexts.ticketAssistantTile.title)}
-                color={themeColor}
-                testID={testID + 'Title'}
-              >
-                {title}
-              </ThemeText>
-            </View>
-
-            <ThemeText
-              type="body__tertiary"
-              color={'secondary'}
-              style={styles.description}
-            >
-              {t(TicketingTexts.ticketAssistantTile.description)}
-            </ThemeText>
-          </View>
-          <TicketMultiple
-            style={styles.illustration}
-            fill={transportThemeSecondaryColor.background}
-            width={theme.icon.size.large}
-            height={theme.icon.size.large}
-          />
-        </TouchableOpacity>
-      )}
-    </View>
-  );
+          <TicketMultiple style={styles.illustration} />
+        </TicketingTile>
+      </View>
+    );
+  }
 };
 
 const useStyles = StyleSheet.createThemeHook((theme) => ({
   tipsAndInformation: {
     flexShrink: 1,
     alignSelf: 'stretch',
-    padding: theme.spacings.xLarge,
-    paddingBottom: theme.spacings.xLarge - 2 * theme.border.width.medium,
-    borderBottomWidth: 2 * theme.border.width.medium,
-    borderRadius: theme.border.radius.small,
-
     marginHorizontal: theme.spacings.medium,
     marginBottom: theme.spacings.large,
   },
-  betaTag: {
-    marginBottom: theme.spacings.xSmall,
-  },
   illustration: {
     marginTop: theme.spacings.small,
-  },
-  description: {marginBottom: theme.spacings.small},
-  iconBox: {
-    backgroundColor: theme.static.status.info.background,
-    display: 'flex',
-    flexDirection: 'row',
-    padding: theme.spacings.xSmall,
-    borderRadius: theme.border.radius.small,
-    marginRight: theme.spacings.xSmall,
-  },
-  titleContainer: {
-    flexDirection: 'row',
-    marginBottom: theme.spacings.small,
-  },
-
-  contentContainer: {
-    flexShrink: 1,
-  },
-  spreadContent: {
-    flex: 1,
-    justifyContent: 'space-between',
   },
 }));
