+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_PurchaseTabScreen/Components/FareProducts/FareProductTile.tsx
@@ -1,25 +1,17 @@
-import {TouchableOpacity, View} from 'react-native';
-import {ThemeText} from '@atb/components/text';
 import React from 'react';
-import {StyleSheet, useTheme} from '@atb/theme';
-import {FareProductIllustration} from './FareProductIllustration';
+
 import {
   FareContractTexts,
-  TicketingTexts,
   useTranslation,
   TranslateFunction,
 } from '@atb/translations';
-import {
-  getStaticColor,
-  getTransportationColor,
-  StaticColor,
-} from '@atb/theme/colors';
 
 import {FareProductTypeConfig} from '@atb/configuration';
 import {useTextForLanguage} from '@atb/translations/utils';
 
 import {useThemeColorForTransportMode} from '@atb/utils/use-transportation-color';
 import {TransportModePair} from '@atb/components/transportation-modes';
+import {TicketingTile} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/TicketingTile';
 
 const modesDisplayLimit = 2;
 
@@ -34,125 +26,38 @@ export const FareProductTile = ({
   testID: string;
   config: FareProductTypeConfig;
 }) => {
-  const styles = useStyles();
   const {t} = useTranslation();
-  const {theme, themeName} = useTheme();
 
   const transportModes = config.transportModes;
-
-  const color: StaticColor = accented ? 'background_accent_3' : 'background_0';
-  const themeColor = getStaticColor(themeName, color);
 
   const transportColor = useThemeColorForTransportMode(
     transportModes[0]?.mode,
     transportModes[0]?.subMode,
   );
 
-  const transportThemePrimaryColor = getTransportationColor(
-    themeName,
-    transportColor,
-    'primary',
-  );
-  const transportThemeSecondaryColor = getTransportationColor(
-    themeName,
-    transportColor,
-    'secondary',
-  );
-
-  const title = useTextForLanguage(config.name);
-  const description = useTextForLanguage(config.description);
-
   const transportModesText = getFareProductTravelModesText(
     transportModes,
     t,
     modesDisplayLimit,
   );
-  const accessibilityLabel = [title, transportModesText, description].join(
-    '. ',
-  );
+
+  const title = useTextForLanguage(config.name) + ', ' + transportModesText;
+  const description = useTextForLanguage(config.description);
+  const accessibilityLabel = [title, description].join('. ');
 
   return (
-    <View
-      style={[
-        styles.fareProduct,
-        {
-          backgroundColor: themeColor.background,
-          borderBottomColor: transportThemePrimaryColor.background,
-        },
-      ]}
+    <TicketingTile
+      accented={accented}
+      onPress={onPress}
       testID={testID}
-    >
-      <TouchableOpacity
-        onPress={onPress}
-        accessible={true}
-        accessibilityLabel={accessibilityLabel}
-        accessibilityHint={t(
-          TicketingTexts.availableFareProducts.navigateToBuy,
-        )}
-        style={styles.spreadContent}
-      >
-        <View style={styles.contentContainer}>
-          <ThemeText
-            type="body__secondary--bold"
-            style={styles.title}
-            accessibilityLabel={title}
-            color={themeColor}
-            testID={testID + 'Title'}
-          >
-            {title + ', ' + transportModesText}
-          </ThemeText>
-          <ThemeText
-            type="body__tertiary"
-            style={styles.description}
-            color={'secondary'}
-          >
-            {description}
-          </ThemeText>
-        </View>
-        <FareProductIllustration
-          style={styles.illustration}
-          config={config}
-          fill={transportThemeSecondaryColor.background}
-          width={theme.icon.size.large}
-          height={theme.icon.size.large}
-        />
-      </TouchableOpacity>
-    </View>
+      config={config}
+      transportColor={transportColor}
+      title={title}
+      description={description}
+      accessibilityLabel={accessibilityLabel}
+    />
   );
 };
-
-const useStyles = StyleSheet.createThemeHook((theme) => ({
-  fareProduct: {
-    width: '100%',
-    flexShrink: 1,
-    alignSelf: 'stretch',
-    marginRight: theme.spacings.medium,
-    padding: theme.spacings.xLarge,
-    paddingBottom: theme.spacings.xLarge - 2 * theme.border.width.medium,
-    borderBottomWidth: 2 * theme.border.width.medium,
-    borderRadius: theme.border.radius.small,
-  },
-  contentContainer: {
-    flexShrink: 1,
-  },
-  iconContainer: {
-    flex: 1,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  spreadContent: {
-    flex: 1,
-    justifyContent: 'space-between',
-  },
-  label: {marginLeft: theme.spacings.xSmall},
-  illustration: {
-    marginTop: theme.spacings.small,
-  },
-  title: {
-    marginBottom: theme.spacings.small,
-  },
-  description: {marginBottom: theme.spacings.small},
-}));
 
 const getFareProductTravelModesText = (
   modes: TransportModePair[],
